requirements.txt
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
colorlog==6.9.0
cryptography==43.0.3
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
mssql-django==1.5
packaging==24.2
pycparser==2.22
PyJWT==2.10.0
pyodbc==5.2.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
wheel==0.44.0
django-filter==25.1
django-cors-headers==4.7.0